clean_lama/requirements.txt
# Core dependencies
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.17.0
scikit-learn>=0.24.0
tqdm>=4.60.0
PyYAML>=5.4.0

# Deep Learning (will be installed by setup.py with correct CUDA version)
# torch>=1.10.0
# torchvision>=0.11.0

# Training
hydra-core>=1.1.0
omegaconf>=2.1.0
pytorch-lightning>=1.5.0
tensorboard>=2.4.0

# Data augmentation
albumentations>=1.0.0

# Model components
kornia>=0.5.0
einops>=0.3.0

# Data loading
webdataset>=0.1.40

# Visualization
matplotlib>=3.3.0
pandas>=1.1.0

# ONNX (will be installed by setup.py based on CUDA availability)
# onnx>=1.10.0
# onnxruntime-gpu>=1.10.0  # or onnxruntime for CPU

# Optional: ONNX optimization
onnx-simplifier>=0.3.6
onnxconverter-common>=1.8.0

# Optional: Better progress bars
rich>=10.0.0

# Optional: Experiment tracking
# wandb>=0.12.0
# mlflow>=1.20.0